requirements.txt
supabase
python-dotenv
balldontlie